+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Integer axes only show integer ticks
 
 ### Fix
+- Remove data attr from class Arc in Python
 - Light improvement of class Arc in Python
 
 ## [0.21.0]